update: babel@7, electron@4, webpack@4
