+++ .babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["@babel/react", "@babel/flow"],
+  "plugins": ["@babel/plugin-transform-flow-strip-types", "@babel/plugin-proposal-class-properties", "babel-plugin-add-module-exports"]
+}
