+++ package.json
@@ -33,6 +33,7 @@
     "postinstall": "electron-builder install-app-deps && node build/downloadDaemon.js"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.2.5",
     "@types/three": "^0.93.1",
     "bluebird": "^3.5.1",
     "breakdance": "^3.0.1",
@@ -44,7 +45,7 @@
     "electron-dl": "^1.11.0",
     "electron-is-dev": "^0.3.0",
     "electron-log": "^2.2.12",
-    "electron-updater": "^2.23.3",
+    "electron-updater": "^4.0.0",
     "electron-window-state": "^4.1.1",
     "find-process": "^1.1.0",
     "formik": "^0.10.4",
@@ -57,6 +58,7 @@
     "mime": "^2.3.1",
     "mixpanel-browser": "^2.17.1",
     "moment": "^2.22.0",
+    "prop-types": "^15.6.2",
     "qrcode.react": "^0.8.0",
     "rc-progress": "^2.0.6",
     "react": "^16.3.0",
@@ -67,9 +69,7 @@
     "react-redux": "^5.0.3",
     "react-simplemde-editor": "^3.6.16",
     "react-toggle": "^4.0.2",
-    "react-transition-group": "1.x",
     "redux": "^3.6.0",
-    "redux-logger": "^3.0.1",
     "redux-persist": "^4.8.0",
     "redux-persist-transform-compress": "^4.2.0",
     "redux-persist-transform-filter": "0.0.16",
@@ -80,31 +80,31 @@
     "render-media": "^3.1.0",
     "reselect": "^3.0.0",
     "semver": "^5.3.0",
-    "shapeshift.io": "^1.3.1",
-    "source-map-support": "^0.5.4",
+    "source-map-support": "^0.5.10",
     "stream-to-blob-url": "^2.1.1",
     "three": "^0.93.0",
-    "tree-kill": "^1.1.0",
     "y18n": "^4.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-transform-flow-strip-types": "^7.2.3",
+    "@babel/preset-flow": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
     "@lbry/color": "^1.0.2",
     "@lbry/components": "^1.5.1",
     "axios": "^0.18.0",
-    "babel-eslint": "^8.2.2",
-    "babel-plugin-module-resolver": "^3.1.1",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.18.0",
+    "babel-eslint": "^9.0.0",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-add-module-exports": "^1.0.0",
     "decompress": "^4.2.0",
     "del": "^3.0.0",
     "devtron": "^1.4.0",
-    "electron": "^2.0.14",
-    "electron-builder": "^20.22.0",
+    "electron": "^4.0.1",
+    "electron-builder": "^20.38.4",
     "electron-devtools-installer": "^2.2.3",
     "electron-publisher-s3": "^20.8.1",
-    "electron-webpack": "^1.13.0",
+    "electron-webpack": "^2.6.2",
     "eslint": "^4.19.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-prettier": "^2.9.0",
@@ -121,12 +121,11 @@
     "i18n-extract": "^0.5.1",
     "json-loader": "^0.5.4",
     "lint-staged": "^7.0.2",
-    "make-runnable": "^1.3.6",
     "node-loader": "^0.6.0",
     "node-sass": "^4.7.2",
     "prettier": "^1.11.1",
     "sass-loader": "^6.0.7",
-    "webpack": "^3.10.0",
+    "webpack": "^4.28.4",
     "webpack-build-notifier": "^0.1.23",
     "yarnhook": "^0.2.0"
   },
