+++ src/main/index.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-// Module imports
+import '@babel/polyfill';
 import keytar from 'keytar';
 import SemVer from 'semver';
 import findProcess from 'find-process';
@@ -279,7 +279,8 @@ process.on('uncaughtException', error => {
 });
 
 // Force single instance application
-const isSecondInstance = app.makeSingleInstance(argv => {
+app.requestSingleInstanceLock();
+app.on('second-instance', (event, argv) => {
   if (rendererWindow) {
     if (
       (process.platform === 'win32' || process.platform === 'linux') &&
@@ -307,7 +308,3 @@ const isSecondInstance = app.makeSingleInstance(argv => {
     rendererWindow.show();
   }
 });
-
-if (isSecondInstance) {
-  app.exit();
-}
