+++ src/renderer/component/viewers/threeViewer/index.jsx
@@ -5,12 +5,53 @@ import classNames from 'classnames';
 import LoadingScreen from 'component/common/loading-screen';
 
 // ThreeJS
-import * as THREE from './internal/three';
+import * as THREE from 'three';
+
 import detectWebGL from './internal/detector';
 import ThreeGrid from './internal/grid';
 import ThreeScene from './internal/scene';
-import ThreeLoader from './internal/loader';
 import ThreeRenderer from './internal/renderer';
+
+global.THREE = THREE;
+
+// Currently it's not possible to import the files within the "examples/js" directory.
+// Fix: https://github.com/mrdoob/three.js/issues/9562#issuecomment-383390251
+// eslint-disable import/no-commonjs
+require('three/examples/js/controls/OrbitControls');
+require('three/examples/js/loaders/LoaderSupport');
+require('three/examples/js/loaders/OBJLoader2');
+require('three/examples/js/loaders/STLLoader');
+// eslint-enable import/no-commonjs
+
+const Manager = ({ onLoad, onStart, onError }) => {
+  const manager = new THREE.LoadingManager();
+  manager.onLoad = onLoad;
+  manager.onStart = onStart;
+  manager.onError = onError;
+
+  return manager;
+};
+
+const Loader = (fileType, manager) => {
+  const fileTypes = {
+    stl: () => new THREE.STLLoader(manager),
+    obj: () => new THREE.OBJLoader2(manager),
+  };
+  return fileTypes[fileType] ? fileTypes[fileType]() : null;
+};
+
+const ThreeLoader = ({ fileType = null, downloadPath = null }, renderModel, managerEvents) => {
+  if (fileType && downloadPath) {
+    const manager = Manager(managerEvents);
+    const loader = Loader(fileType, manager);
+
+    if (loader) {
+      loader.load(`file://${downloadPath}`, data => {
+        renderModel(fileType, data);
+      });
+    }
+  }
+};
 
 type viewerTheme = {
   gridColor: string,
