+++ src/renderer/component/viewers/threeViewer/internal/grid.js
@@ -1,4 +1,4 @@
-import { GridHelper, Color } from './three';
+import { GridHelper, Color } from 'three';
 
 const ThreeGrid = ({ size, gridColor, centerLineColor }) => {
   const divisions = size / 2;
