+++ src/renderer/constants/icons.js
@@ -25,6 +25,7 @@ export const PHONE = 'Phone';
 export const CHECK = 'CheckCircle';
 export const HEART = 'Heart';
 export const UNLOCK = 'Unlock';
+export const LOCK = 'Lock';
 export const CHECK_SIMPLE = 'Check';
 export const GLOBE = 'Globe';
 export const EXTERNAL_LINK = 'ExternalLink';
@@ -36,3 +37,4 @@ export const TWITTER = 'Twitter';
 export const CREDIT_CARD = 'CreditCard';
 export const SETTINGS = 'Settings';
 export const USERS = 'Users';
+export const MENU = 'Menu';
